Move stores to separate module
